requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain_text_splitters
langchain-openai
chromadb
pypdf
python-dotenv
gradio
pytest
langchain-chroma